+++ packages/plugin-gallery/web/src/components/gallery-controls/thumbnail-placement-selector.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
 import { mergeStyles } from 'wix-rich-content-common';
 import { SelectionList, SelectionListItem } from 'wix-rich-content-ui-components';
 import styles from '../../../statics/styles/thumbnail-placement-selector.rtlignore.scss';
@@ -36,11 +35,7 @@ class ThumbnailPlacementSelector extends Component {
     ) : (
       <div className={this.styles.thumbnailPlacementSelector_tile}>
         <LayoutThumbnailsIcon
-          className={classNames(
-            this.styles.thumbnailPlacementSelector_icon,
-            this.styles[`thumbnailPlacementSelector_${item.alignment}`],
-            { [this.styles.thumbnailPlacementSelector_icon_selected]: selected }
-          )}
+          className={this.styles[`thumbnailPlacementSelector_${item.alignment}`]}
         />
       </div>
     );
@@ -58,7 +53,7 @@ class ThumbnailPlacementSelector extends Component {
           className={this.styles.thumbnailPlacementSelector_grid}
           dataSource={this.dataSource}
           dataMapper={this.dataMapper}
-          renderItem={data => this.renderOption(data)}
+          renderItem={this.renderOption}
           value={value}
           onChange={onChange}
           optionClassName={styles.thumbnailPlacementSelector_option}